+++ seldom/request.py
@@ -9,6 +9,7 @@ IMG = ["jpg", "jpeg", "gif", "bmp", "webp"]
 def request(func):
     def wrapper(*args, **kwargs):
         func_name = func.__name__
+        print("\n")
         log.info('-------------- Request -----------------[🚀]')
         try:
             url = list(args)[1]
@@ -47,18 +48,19 @@ def request(func):
 
         ResponseResult.status_code = r.status_code
         log.info("-------------- Response ----------------[🛬️]")
+        resp_time = r.elapsed.total_seconds()
         try:
             resp = r.json()
-            log.debug(f"[type]: json \n")
+            log.debug(f"[type]: json      [time]: {resp_time} \n")
             log.debug(f"[response]:\n {resp} \n")
             ResponseResult.response = resp
         except BaseException as msg:
-            log.debug("[warning]: {} \n".format(msg))
+            log.debug(f"[warning]: {msg} \n")
             if img_file is True:
-                log.debug("[type]: {}".format(file_type))
+                log.debug(f"[type]: {file_type}      [time]: {resp_time}")
                 ResponseResult.response = r.content
             else:
-                log.debug("[type]: text \n")
+                log.debug(f"[type]: text      [time]: {resp_time}\n")
                 log.debug(f"[response]:\n {r.text} \n")
                 ResponseResult.response = r.text
 
@@ -104,20 +106,58 @@ class HttpRequest(object):
         """
         return ResponseResult.response
 
-    @property
-    def session(self):
-        """
-        A Requests session.
-        """
-        s = requests.Session()
-        return s
+    class Session(requests.Session):
 
-    @staticmethod
-    def request(method=None, url=None, headers=None, files=None, data=None,
-                params=None, auth=None, cookies=None, hooks=None, json=None):
-        """
-        A user-created :class:`Request <Request>` object.
-        """
-        req = requests.Request(method, url, headers, files, data,
-                               params, auth, cookies, hooks, json)
-        return req
+        @request
+        def get(self, url, **kwargs):
+            r"""Sends a GET request. Returns :class:`Response` object.
+
+            :param url: URL for the new :class:`Request` object.
+            :param \*\*kwargs: Optional arguments that ``request`` takes.
+            :rtype: requests.Response
+            """
+            if (Seldom.base_url is not None) and ("http" not in url):
+                url = Seldom.base_url + url
+            kwargs.setdefault('allow_redirects', True)
+            return self.request('GET', url, **kwargs)
+
+        @request
+        def post(self, url, data=None, json=None, **kwargs):
+            r"""Sends a POST request. Returns :class:`Response` object.
+
+            :param url: URL for the new :class:`Request` object.
+            :param data: (optional) Dictionary, list of tuples, bytes, or file-like
+                object to send in the body of the :class:`Request`.
+            :param json: (optional) json to send in the body of the :class:`Request`.
+            :param \*\*kwargs: Optional arguments that ``request`` takes.
+            :rtype: requests.Response
+            """
+            if (Seldom.base_url is not None) and ("http" not in url):
+                url = Seldom.base_url + url
+            return self.request('POST', url, data=data, json=json, **kwargs)
+
+        @request
+        def put(self, url, data=None, **kwargs):
+            r"""Sends a PUT request. Returns :class:`Response` object.
+
+            :param url: URL for the new :class:`Request` object.
+            :param data: (optional) Dictionary, list of tuples, bytes, or file-like
+                object to send in the body of the :class:`Request`.
+            :param \*\*kwargs: Optional arguments that ``request`` takes.
+            :rtype: requests.Response
+            """
+            if (Seldom.base_url is not None) and ("http" not in url):
+                url = Seldom.base_url + url
+            return self.request('PUT', url, data=data, **kwargs)
+
+        @request
+        def delete(self, url, **kwargs):
+            r"""Sends a DELETE request. Returns :class:`Response` object.
+
+            :param url: URL for the new :class:`Request` object.
+            :param \*\*kwargs: Optional arguments that ``request`` takes.
+            :rtype: requests.Response
+            """
+            if (Seldom.base_url is not None) and ("http" not in url):
+                url = Seldom.base_url + url
+            return self.request('DELETE', url, **kwargs)