+++ CHANGES.md
@@ -1,3 +1,25 @@
+### 2.7.0
+
+* 功能：引入`loguru` 库用于打印日志（之前使用python默认logging总有一些重复打印或不打印的问题）。
+* 功能：web自动化增加一套方法链（method chaining）的API，[文档](./docs/chaining.md)。
+* 功能：支持手动指定浏览器驱动路径，[文档](./docs/other.md)。
+
+### 2.6.0
+
+* 移除：自带的`HTMLTestRunner`，HTML报告采用`XTestRunner`。
+* 移除：对`unittest-xml-reporting`库的依赖，XML报告使用`XTestRunner`。
+* 修改：`SMTP`类发送邮件方法 `sender()` -> `sendmail()`， 发送邮件样式采用`XTestRunner` [文档](./docs/advanced.md)。
+* 增加：`seldom.main()`方法增加`tester` 参数，用于设置测试人员名字，默认`Anonymous` [文档](./docs/quick_start.md)。
+* 增加：`seldom.main()`方法增加`language` 参数，用于设置报告中英文`en/zh-CN`，默认`en` [文档](./docs/quick_start.md)。
+* 增加：发送钉钉功能 [文档](./docs/advanced.md) 。
+* 修改：接口测试 `self.session` -> `self.Session()` [文档](./docs/http.md)。  
+* 移除：接口测试 `self.request()` 方法移除（注：该方法原本不可用）。
+
+### 2.5.1
+* 功能：Http接口测试使用日志打印接口信息
+* 功能：Http接口测试打印`json`参数 [83](https://github.com/SeldomQA/seldom/issues/83)
+* 修复：Web UI测试`self.Key()` 无法定位元素的问题
+
 ### 2.5.0
 * 功能：支持测试平台化 [参考](https://github.com/SeldomQA/seldom/blob/master/docs/platform.md) 。
 * 功能：utils 增加`file`类，获取当前文件目录更方便。